flask/requirements.txt
flask
kafka-python
prometheus_client